Vietnamese translation
Dear Adrian(sorry if your name is not adrian)
I'm a Vietnamese student who stumbled across this terrific game. But i see that the game doesn't have a Vietnamese version and well, my friends suck at English, so I spent my evening translating all of the game materials for ya. If there is anything to add to the game, you should consider for the spy to have a early guess, and the others to have a vote system. Have a good day! 
-Tam-